Changes 2022-01-01

+++ src/AnglePlatform.java
@@ -2,7 +2,7 @@ package src;
 
 public class AnglePlatform implements IPlatform {
 
-    public double platformAngle = 0.0;  // Start value
+    public int platformAngle = 0;  // Start value
     public int platformCapacity;
     public int loadOnPlatform;
 
@@ -10,29 +10,29 @@ public class AnglePlatform implements IPlatform {
         platformCapacity will be set to received argument.
         Argument must be positive. */
     public AnglePlatform(int platformCapacity){ 
-        this.platformAngle = 0.0;
+        this.platformAngle = 0;
         this.platformCapacity = platformCapacity;
     }
 
     /** Call this method to raise platformAngle.
         Platform can not be raised more than 70 degrees. */
     public void raise(){
-        if (this.platformAngle < 70.0){
+        if (this.platformAngle < 70){
             this.platformAngle += 1;
         }
-        if (this.platformAngle > 70.0){
-            this.platformAngle = 70.0;
+        if (this.platformAngle > 70){
+            this.platformAngle = 70;
         }
 
     }
     /** Call this method to lower the platformAngle.
         Platform can not be lowered more than 0 degrees */
     public void lower(){
-        if (this.platformAngle > 0.0){
+        if (this.platformAngle > 0){
             this.platformAngle -= 1;
         }
-        if (this.platformAngle < 0.0){
-            this.platformAngle = 0.0;
+        if (this.platformAngle < 0){
+            this.platformAngle = 0;
         }
 
     }
